+++ tests/Get-DbaStartupParameter.Tests.ps1
@@ -16,8 +16,12 @@ Describe "$CommandName Unit Tests" -Tag 'UnitTests' {
         }
     }
 }
-<#
-    Integration test should appear below and are custom to the command you are writing.
-    Read https://github.com/sqlcollaborative/dbatools/blob/development/contributing.md#tests
-    for more guidence.
-#>
+
+Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
+    Context "Command actually works" {
+        $results = Get-DbaStartupParameter -SqlInstance $script:Instance2
+        it "Gets Results" {
+            $results | Should Not Be $null
+        }
+    }
+}